[GTK] Test /webkit2/WebKitWebView/geolocation-permission-requests is failing since r201423 https://bugs.webkit.org/show_bug.cgi?id=158200
Reviewed by Philippe Normand.

This is because geolocation is no longer allowed for non secure sites, like HTTP. In that case
POSITION_UNAVAILABLE is returned without asking the API layer.

* TestWebKitAPI/Tests/WebKit2Gtk/TestUIClient.cpp:
(testWebViewGeolocationPermissionRequests): Keep the HTTP case to check that it indeed returns
POSITION_UNAVAILABLE and use HTTPS URLs to check permission requests are allowed or denied. Also stop using the
document title, and use user script messages that are more reliable instead.

Identifier: 176305@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@201505 268f45cc-cd09-0410-ab3c-d52691b4dbfc