Fix round-down goof in Air.js's ShuffleCustom.forEachArg https://bugs.webkit.org/show_bug.cgi?id=158674
Reviewed by Michael Saboff.

x / 3 * 3 is not how you round down to multiples of 3 in JavaScript. You need to do
Math.floor(x / 3) * 3 instead.

This is a benign change, because having extra arguments to a Shuffle is not something we
actually take advantage of yet. But I think it's best to match the original C++ code's
intent.

* Air.js/custom.js:
(const.ShuffleCustom.forEachArg):



Identifier: 176775@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@201996 268f45cc-cd09-0410-ab3c-d52691b4dbfc